rename components for improved clarity and consistency
