+++ frontend/src/app/components/LessonCard.module.css
