+++ frontend/src/app/components/LessonCard.tsx
@@ -1,16 +1,8 @@
-import styles from "./TeacherViewCard.module.css";
+import styles from "./LessonCard.module.css";
 import Link from "next/link";
 import { Lesson } from "../Types";
 
-interface TeacherViewCardProps {
-  lesson: Lesson;
-  studentId: string;
-}
-
-export default function TeacherViewCard({
-  lesson,
-  studentId,
-}: TeacherViewCardProps) {
+export default function LessonCard({ lesson }: { lesson: Lesson }) {
   const { title, status, id } = lesson;
   return (
     <div className={styles.card}>
@@ -18,10 +10,7 @@ export default function TeacherViewCard({
         <h3 className={styles.title}>{title}</h3>
         <h3 className={styles.status}>{status}</h3>
       </div>
-      <Link
-        href={`/teacher/student/${studentId}/lesson/${id}`}
-        className={styles.button}
-      >
+      <Link href={`/lessons/${id}`} className={styles.button}>
         Go to Lesson
       </Link>
     </div>
