+++ frontend/src/app/components/StudentLessons.tsx
@@ -1,6 +1,6 @@
 "use client";
 import styles from "./StudentLessons.module.css";
-import Card from "../components/Card";
+import LessonCard from "../components/LessonCard";
 import { Lesson } from "../Types";
 import { useSWRAxios } from "../hooks/useSWRAxios";
 import { API_PATHS } from "../constants/apiKeys";
@@ -12,7 +12,7 @@ export default function StudentLessons() {
     <div className={styles["cards-container"]}>
       {lessons &&
         lessons.map((lesson) => (
-          <Card key={lesson.title} lesson={lesson} />
+          <LessonCard key={lesson.title} lesson={lesson} />
         ))}
     </div>
   );
