+++ frontend/src/app/components/TeacherLessonCard.module.css
