+++ frontend/src/app/components/TeacherLessonCard.tsx
@@ -1,8 +1,16 @@
-import styles from "./Card.module.css";
+import styles from "./TeacherLessonCard.module.css";
 import Link from "next/link";
 import { Lesson } from "../Types";
 
-export default function Card({ lesson }: { lesson: Lesson }) {
+interface TeacherLessonCardProps {
+  lesson: Lesson;
+  studentId: string;
+}
+
+export default function TeacherLessonCard({
+  lesson,
+  studentId,
+}: TeacherLessonCardProps) {
   const { title, status, id } = lesson;
   return (
     <div className={styles.card}>
@@ -10,7 +18,10 @@ export default function Card({ lesson }: { lesson: Lesson }) {
         <h3 className={styles.title}>{title}</h3>
         <h3 className={styles.status}>{status}</h3>
       </div>
-      <Link href={`/lessons/${id}`} className={styles.button}>
+      <Link
+        href={`/teacher/student/${studentId}/lesson/${id}`}
+        className={styles.button}
+      >
         Go to Lesson
       </Link>
     </div>
