+++ frontend/src/app/components/TeacherLessonDetails.module.css
