+++ frontend/src/app/components/TeacherLessonsList.tsx
@@ -1,11 +1,11 @@
 "use client";
 import { useSWRAxios } from "../hooks/useSWRAxios";
 import { Lesson } from "../Types";
-import TeacherViewCard from "./TeacherViewCard";
-import styles from "./TeacherViewLessons.module.css";
+import TeacherLessonCard from "./TeacherLessonCard";
+import styles from "./TeacherLessonsList.module.css";
 import { API_PATHS } from "../constants/apiKeys";
 
-export default function TeacherViewLessons({ id }: { id: string }) {
+export default function TeacherLessonsList({ id }: { id: string }) {
   const { data: lessons } = useSWRAxios<Lesson[]>(
     API_PATHS.TEACHER_STUDENT_LESSONS(id)
   );
@@ -13,7 +13,7 @@ export default function TeacherViewLessons({ id }: { id: string }) {
     <section className={styles.container}>
       {lessons &&
         lessons.map((lesson: Lesson) => (
-          <TeacherViewCard
+          <TeacherLessonCard
             key={lesson.id}
             lesson={lesson}
             studentId={id}
