+++ frontend/src/app/teacher/student/[studentId]/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import TeacherViewLessons from "@/app/components/TeacherViewLessons";
+import TeacherLessonsList from "@/app/components/TeacherLessonsList";
 import { ErrorBoundary } from "react-error-boundary";
 import { Suspense, use } from "react";
 import StudentInfo from "@/app/components/StudentInfo";
@@ -20,7 +20,7 @@ export default function StudentPage({ params }: StudentPageProps) {
     >
       <Suspense fallback={<div>Loading student details...</div>}>
         <StudentInfo id={id} />
-        <TeacherViewLessons id={id} />
+        <TeacherLessonsList id={id} />
       </Suspense>
     </ErrorBoundary>
   );
